requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
jpype1==1.5.0
